sarc: Write data alignment to yaz0 header (when needed)
When no alignment requirement is passed to the decompressor load
function, max(yaz0_header->alignment, 0x20) will be used as the
alignment value for the read buffer.

It seems that BotW always passes on the alignment requirements
to the decompressor but let's not rely on this. This commit
changes the sarc code to write the alignment to the yaz0 header.